+++ packages/clerk-js/src/ui/components/UserProfile/DeleteUserForm.tsx
@@ -1,18 +1,21 @@
-import { useUser } from '@clerk/shared/react';
+import { useClerk, useUser } from '@clerk/shared/react';
 
 import { useSignOutContext } from '../../contexts';
 import { Col, localizationKeys, Text, useLocalizations } from '../../customizables';
 import type { FormProps } from '../../elements';
 import { Form, FormButtons, FormContainer, useCardState, withCardStateProvider } from '../../elements';
+import { useMultipleSessions } from '../../hooks/useMultipleSessions';
 import { handleError, useFormControl } from '../../utils';
 
 type DeleteUserFormProps = FormProps;
 export const DeleteUserForm = withCardStateProvider((props: DeleteUserFormProps) => {
   const { onReset } = props;
   const card = useCardState();
-  const { navigateAfterSignOut } = useSignOutContext();
+  const { navigateAfterSignOut, navigateAfterMultiSessionSingleSignOutUrl } = useSignOutContext();
   const { user } = useUser();
   const { t } = useLocalizations();
+  const { otherSessions } = useMultipleSessions({ user });
+  const { setActive } = useClerk();
 
   const confirmationField = useFormControl('deleteConfirmation', '', {
     type: 'text',
@@ -36,7 +39,13 @@ export const DeleteUserForm = withCardStateProvider((props: DeleteUserFormProps)
       }
 
       await user.delete();
-      await navigateAfterSignOut();
+      // TODO: Investigate if we need to call `setActive` with {session: null}
+      const navigationCallback =
+        otherSessions.length === 0 ? navigateAfterSignOut : navigateAfterMultiSessionSingleSignOutUrl;
+      return await setActive({
+        session: null,
+        beforeEmit: navigationCallback,
+      });
     } catch (e) {
       handleError(e, [], card.setError);
     }