+++ packages/elements/src/internals/machines/sign-in/router.machine.ts
@@ -309,7 +309,7 @@ export const SignInRouterMachine = setup({
       ],
     },
     Start: {
-      tags: 'route:start',
+      tags: ['step:start'],
       exit: 'clearFormErrors',
       invoke: {
         id: 'start',
@@ -358,7 +358,7 @@ export const SignInRouterMachine = setup({
       },
     },
     FirstFactor: {
-      tags: 'route:first-factor',
+      tags: ['step:first-factor', 'step:verifications'],
       invoke: {
         id: 'firstFactor',
         src: 'firstFactorMachine',
@@ -414,7 +414,7 @@ export const SignInRouterMachine = setup({
           },
         },
         ChoosingStrategy: {
-          tags: ['route:choose-strategy'],
+          tags: ['step:choose-strategy'],
           on: {
             'NAVIGATE.PREVIOUS': {
               description: 'Go to Idle, and also tell firstFactor to go to Pending',
@@ -424,7 +424,7 @@ export const SignInRouterMachine = setup({
           },
         },
         ForgotPassword: {
-          tags: ['route:forgot-password'],
+          tags: ['step:forgot-password'],
           on: {
             'NAVIGATE.PREVIOUS': 'Idle',
           },
@@ -432,7 +432,7 @@ export const SignInRouterMachine = setup({
       },
     },
     SecondFactor: {
-      tags: 'route:second-factor',
+      tags: ['step:second-factor', 'step:verifications'],
       invoke: {
         id: 'secondFactor',
         src: 'secondFactorMachine',
@@ -460,10 +460,37 @@ export const SignInRouterMachine = setup({
             target: 'ResetPassword',
           },
         ],
+        'STRATEGY.UPDATE': {
+          description: 'Send event to verification machine to update the current strategy.',
+          actions: sendTo('secondFactor', ({ event }) => event),
+          target: '.Idle',
+        },
+      },
+      initial: 'Idle',
+      states: {
+        Idle: {
+          on: {
+            'NAVIGATE.CHOOSE_STRATEGY': {
+              description: 'Navigate to choose strategy screen.',
+              actions: sendTo('secondFactor', ({ event }) => event),
+              target: 'ChoosingStrategy',
+            },
+          },
+        },
+        ChoosingStrategy: {
+          tags: ['step:choose-strategy'],
+          on: {
+            'NAVIGATE.PREVIOUS': {
+              description: 'Go to Idle, and also tell firstFactor to go to Pending',
+              target: 'Idle',
+              actions: sendTo('secondFactor', { type: 'NAVIGATE.PREVIOUS' }),
+            },
+          },
+        },
       },
     },
     ResetPassword: {
-      tags: 'route:reset-password',
+      tags: ['step:reset-password'],
       invoke: {
         id: 'resetPassword',
         src: 'resetPasswordMachine',
@@ -499,7 +526,7 @@ export const SignInRouterMachine = setup({
       },
     },
     Callback: {
-      tags: 'route:callback',
+      tags: ['step:callback'],
       entry: sendTo(ThirdPartyMachineId, { type: 'CALLBACK' }),
       on: {
         NEXT: [
@@ -535,7 +562,7 @@ export const SignInRouterMachine = setup({
       },
     },
     Error: {
-      tags: 'route:error',
+      tags: ['step:error'],
       on: {
         NEXT: {
           target: 'Start',