+++ packages/elements/src/internals/machines/sign-in/verification.machine.ts
@@ -1,12 +1,13 @@
+import { isClerkAPIResponseError } from '@clerk/shared/error';
 import type {
   AttemptFirstFactorParams,
   EmailCodeAttempt,
   PasswordAttempt,
   PhoneCodeAttempt,
   PrepareFirstFactorParams,
+  PrepareSecondFactorParams,
   ResetPasswordEmailCodeAttempt,
   ResetPasswordPhoneCodeAttempt,
-  SignInFactor,
   SignInFirstFactor,
   SignInResource,
   SignInSecondFactor,
@@ -18,7 +19,7 @@ import { assign, fromPromise, log, sendTo, setup } from 'xstate';
 import { RESENDABLE_COUNTDOWN_DEFAULT } from '~/internals/constants';
 import { ClerkElementsRuntimeError } from '~/internals/errors';
 import type { FormFields } from '~/internals/machines/form';
-import type { WithParams } from '~/internals/machines/shared';
+import type { SignInStrategyName, WithParams } from '~/internals/machines/shared';
 import { sendToLoading } from '~/internals/machines/shared';
 import { determineStartingSignInFactor, determineStartingSignInSecondFactor } from '~/internals/machines/sign-in/utils';
 import { assertActorEventError, assertIsDefined } from '~/internals/machines/utils/assert';
@@ -34,11 +35,11 @@ export type DetermineStartingFactorInput = {
   parent: SignInRouterMachineActorRef;
 };
 
-export type PrepareFirstFactorInput = WithParams<SignInFirstFactor | null> & {
+export type PrepareFirstFactorInput = WithParams<PrepareFirstFactorParams> & {
   parent: SignInRouterMachineActorRef;
   resendable: boolean;
 };
-export type PrepareSecondFactorInput = WithParams<SignInSecondFactor | null> & {
+export type PrepareSecondFactorInput = WithParams<PrepareSecondFactorParams> & {
   parent: SignInRouterMachineActorRef;
   resendable: boolean;
 };
@@ -66,8 +67,8 @@ export const SignInVerificationMachineId = 'SignInVerification';
 
 const SignInVerificationMachine = setup({
   actors: {
-    determineStartingFactor: fromPromise<SignInFactor | null, DetermineStartingFactorInput>(() =>
-      Promise.reject(new ClerkElementsRuntimeError('Actor `determineStartingFactor` must be overridden')),
+    determineStartingFactor: fromPromise<SignInFirstFactor | SignInSecondFactor | null, DetermineStartingFactorInput>(
+      () => Promise.reject(new ClerkElementsRuntimeError('Actor `determineStartingFactor` must be overridden')),
     ),
     prepare: fromPromise<SignInResource, PrepareFirstFactorInput | PrepareSecondFactorInput>(() =>
       Promise.reject(new ClerkElementsRuntimeError('Actor `prepare` must be overridden')),
@@ -92,49 +93,52 @@ const SignInVerificationMachine = setup({
         return;
       }
 
-      if (
-        !clerk.client.signIn.supportedFirstFactors.every(factor =>
-          context.registeredStrategies.has(factor.strategy as unknown as SignInFactor),
-        )
-      ) {
-        console.warn(
-          `Clerk: Your instance is configured to support these strategies: ${clerk.client.signIn.supportedFirstFactors
-            .map((factor: any) => factor.strategy)
-            .join(', ')}, but the rendered strategies are: ${Array.from(context.registeredStrategies).join(
-            ', ',
-          )}. Before deploying your app, make sure to render a <Strategy> component for each supported strategy. For more information, visit the documentation: https://clerk.com/docs/elements/reference/sign-in#strategy`,
-        );
-      }
+      // Only show these warnings in development!
+      if (process.env.NODE_ENV === 'development') {
+        if (
+          !clerk.client.signIn.supportedFirstFactors.every(factor => context.registeredStrategies.has(factor.strategy))
+        ) {
+          console.warn(
+            `Clerk: Your instance is configured to support these strategies: ${clerk.client.signIn.supportedFirstFactors
+              .map(factor => factor.strategy)
+              .join(', ')}, but the rendered strategies are: ${Array.from(context.registeredStrategies).join(
+              ', ',
+            )}. Make sure to render a <Strategy> component for each supported strategy. More information: https://clerk.com/docs/elements/reference/sign-in#strategy`,
+          );
+        }
 
-      if (
-        clerk.client.signIn.supportedSecondFactors &&
-        !clerk.client.signIn.supportedSecondFactors.every(factor =>
-          context.registeredStrategies.has(factor.strategy as unknown as SignInFactor),
-        )
-      ) {
-        console.warn(
-          `Clerk: Your instance is configured to support these 2FA strategies: ${[
-            ...clerk.client.signIn.supportedSecondFactors,
-          ]
-            .map(f => f.strategy)
-            .join(', ')}, but the rendered strategies are: ${Array.from(context.registeredStrategies).join(
-            ', ',
-          )}. Before deploying your app, make sure to render a <Strategy> component for each supported strategy. For more information, visit the documentation: https://clerk.com/docs/elements/reference/sign-in#strategy`,
-        );
-      }
+        if (
+          clerk.client.signIn.supportedSecondFactors &&
+          !clerk.client.signIn.supportedSecondFactors.every(factor => context.registeredStrategies.has(factor.strategy))
+        ) {
+          console.warn(
+            `Clerk: Your instance is configured to support these 2FA strategies: ${clerk.client.signIn.supportedSecondFactors
+              .map(f => f.strategy)
+              .join(', ')}, but the rendered strategies are: ${Array.from(context.registeredStrategies).join(
+              ', ',
+            )}. Make sure to render a <Strategy> component for each supported strategy. More information: https://clerk.com/docs/elements/reference/sign-in#strategy`,
+          );
+        }
 
-      if (
-        process.env.NODE_ENV === 'development' &&
-        context.currentFactor?.strategy &&
-        !context.registeredStrategies.has(context.currentFactor?.strategy as unknown as SignInFactor)
-      ) {
-        throw new ClerkElementsRuntimeError(
-          `Your sign-in attempt is missing a ${context.currentFactor?.strategy} strategy. Make sure <Strategy name="${context.currentFactor?.strategy}"> is rendered in your flow. For more information, visit the documentation: https://clerk.com/docs/elements/reference/sign-in#strategy`,
+        const strategiesUsedButNotActivated = Array.from(context.registeredStrategies).filter(
+          strategy => !clerk.client.signIn.supportedFirstFactors.some(supported => supported.strategy === strategy),
         );
-      } else if (process.env.NODE_ENV === 'development' && !context.currentFactor?.strategy) {
-        throw new ClerkElementsRuntimeError(
-          'Unable to determine an authentication strategy to verify. This means your instance is misconfigured. Visit the Clerk Dashboard and verify that your instance has authentication strategies enabled: https://dashboard.clerk.com/last-active?path=/user-authentication/email-phone-username',
-        );
+
+        if (strategiesUsedButNotActivated.length > 0) {
+          console.warn(
+            `Clerk: These rendered strategies are not configured for your instance: ${strategiesUsedButNotActivated.join(', ')}. If this is unexpected, make sure to enable them in your Clerk dashboard: https://dashboard.clerk.com/last-active?path=/user-authentication/email-phone-username`,
+          );
+        }
+
+        if (context.currentFactor?.strategy && !context.registeredStrategies.has(context.currentFactor?.strategy)) {
+          throw new ClerkElementsRuntimeError(
+            `Your sign-in attempt is missing a ${context.currentFactor?.strategy} strategy. Make sure <Strategy name="${context.currentFactor?.strategy}"> is rendered in your flow. More information: https://clerk.com/docs/elements/reference/sign-in#strategy`,
+          );
+        } else if (!context.currentFactor?.strategy) {
+          throw new ClerkElementsRuntimeError(
+            'Unable to determine an authentication strategy to verify. This means your instance is misconfigured. Visit the Clerk Dashboard and verify that your instance has authentication strategies enabled: https://dashboard.clerk.com/last-active?path=/user-authentication/email-phone-username',
+          );
+        }
       }
     },
     sendToNext: ({ context, event }) =>
@@ -151,14 +155,17 @@ const SignInVerificationMachine = setup({
       },
     ),
     setConsoleError: ({ event }) => {
-      if (process.env.NODE_ENV === 'development') {
-        assertActorEventError(event);
-
-        throw new ClerkElementsRuntimeError(`Unable to fulfill the prepare or attempt request for the sign-in verification.
-Error: ${event.error.message}
-
-Please open an issue if you continue to run into this issue.`);
+      if (process.env.NODE_ENV !== 'development') {
+        return;
       }
+
+      assertActorEventError(event);
+
+      const error = isClerkAPIResponseError(event.error) ? event.error.errors[0].longMessage : event.error.message;
+
+      console.error(`Unable to fulfill the prepare or attempt request for the sign-in verification.
+      Error: ${error}
+      Please open an issue if you continue to run into this issue.`);
     },
   },
   guards: {
@@ -174,7 +181,7 @@ Please open an issue if you continue to run into this issue.`);
     formRef: input.formRef,
     loadingStep: 'verifications',
     parent: input.parent,
-    registeredStrategies: new Set<SignInFactor>(),
+    registeredStrategies: new Set<SignInStrategyName>(),
     resendable: false,
     resendableAfter: RESENDABLE_COUNTDOWN_DEFAULT,
   }),
@@ -229,7 +236,7 @@ Please open an issue if you continue to run into this issue.`);
         input: ({ context }) => ({
           parent: context.parent,
           resendable: context.resendable,
-          params: context.currentFactor as SignInFirstFactor | null,
+          params: context.currentFactor as PrepareFirstFactorParams,
         }),
         onDone: {
           actions: 'resendableReset',
@@ -327,7 +334,7 @@ Please open an issue if you continue to run into this issue.`);
         src: 'attempt',
         input: ({ context }) => ({
           parent: context.parent,
-          currentFactor: context.currentFactor as SignInFirstFactor,
+          currentFactor: context.currentFactor as SignInFirstFactor | null,
           fields: context.formRef.getSnapshot().context.fields,
         }),
         onDone: {
@@ -359,7 +366,7 @@ export const SignInFirstFactorMachine = SignInVerificationMachine.provide({
       );
     }),
     prepare: fromPromise(async ({ input }) => {
-      const { params, parent, resendable } = input;
+      const { params, parent, resendable } = input as PrepareFirstFactorInput;
       const clerk = parent.getSnapshot().context.clerk;
 
       // If a prepare call has already been fired recently, don't re-send
@@ -372,7 +379,7 @@ export const SignInFirstFactorMachine = SignInVerificationMachine.provide({
 
       assertIsDefined(params, 'First factor params');
 
-      return await clerk.client.signIn.prepareFirstFactor(params as PrepareFirstFactorParams);
+      return await clerk.client.signIn.prepareFirstFactor(params);
     }),
     attempt: fromPromise(async ({ input }) => {
       const { currentFactor, fields, parent } = input as AttemptFirstFactorInput;